requirements.txt
# requirements.txt
streamlit==1.31.0
anthropic==0.8.1
pillow==10.2.0
python-dotenv==1.0.0
fpdf2==2.7.6
